requirements.txt
argh==0.26.2
MySQL-python==1.2.5
pathtools==0.1.2
PyYAML==3.12
watchdog==0.8.3
